v4: deleted /test route and added all content to / route
